refactor: Split test files into multiple for better organization
