--- editor/server/test/services/test_transactions.py
@@ -1,1550 +1,0 @@
-import copy
-from pathlib import Path
-from typing import cast
-
-import pytest
-from pydantic import ValidationError
-
-from editor.models import FlType, Program, QualifiedID, elements
-from editor.models.edit_errors import BadEdit
-from editor.models.elements import find_element
-from editor.services.intelligence import (
-    IntelligenceReadInterface,
-    IntelligenceService,
-)
-from editor.services.module_editor import ModuleEditor
-from editor.services.transaction import (
-    AddComment,
-    AddConduit,
-    AddDecl,
-    AddDeclInterface,
-    AddNode,
-    ConstantParams,
-    DeclParameterParams,
-    DeclReturnParams,
-    FunctionParams,
-    MoveElement,
-    OperatorParams,
-    RemoveItem,
-    RenameDeclaration,
-    ResizeElement,
-    UpdateComment,
-    UpdateConstant,
-    UpdateFuncParam,
-    UpdateOperator,
-)
-from editor.services.transaction.add_node import CallParams
-
-
-@pytest.fixture
-def basic_program() -> Program:
-    return Program(
-        [
-            elements.Function(
-                name="foo",
-                location=elements.Location(10, 10, 3, 100, 100),
-                id=1,
-            ),
-            elements.Function(
-                name="bar",
-                location=elements.Location(210, 10, 2, 100, 100),
-                id=2,
-                nodes=[
-                    elements.BinaryOperator(
-                        id=1,
-                        location=elements.Location(15, 2, 1, 5, 5),
-                        op=elements.Operator.PLUS,
-                    ),
-                    elements.Constant(
-                        id=2,
-                        location=elements.Location(2, 2, 1, 5, 5),
-                        value="3.0",
-                        flType=FlType.F64,
-                    ),
-                    elements.Constant(
-                        id=3,
-                        location=elements.Location(2, 12, 1, 5, 5),
-                        value="2.0",
-                        flType=FlType.F64,
-                    ),
-                ],
-                conduits=[
-                    elements.Conduit(
-                        id=5,
-                        input=2,
-                        children=[elements.Conduit.Output(target=1, index=1)],
-                    )
-                ],
-            ),
-            elements.Function(
-                name="qux",
-                location=elements.Location(210, 10, 2, 100, 110),
-                id=3,
-                nodes=[
-                    elements.UnaryOperator(
-                        id=2,
-                        location=elements.Location(15, 2, 1, 5, 5),
-                        op=elements.Operator.MINUS,
-                    ),
-                    elements.Constant(
-                        id=3,
-                        location=elements.Location(2, 12, 1, 5, 5),
-                        value="2.0",
-                        flType=FlType.F64,
-                    ),
-                ],
-                conduits=[
-                    elements.Conduit(
-                        id=5,
-                        input=3,
-                        children=[elements.Conduit.Output(target=2, index=0)],
-                    )
-                ],
-            ),
-            elements.Function(
-                name="xyzzy",
-                location=elements.Location(410, 10, 2, 100, 100),
-                id=4,
-                inputs=[
-                    elements.Parameter(name="a", id=1, flType=FlType.I32),
-                    elements.Parameter(name="b", id=2, flType=FlType.I32),
-                ],
-                outputs=[
-                    elements.Return(id=3, flType=FlType.I32),
-                ],
-            ),
-        ]
-    )
-
-
-@pytest.fixture
-def editor(basic_program: Program) -> ModuleEditor:
-    editor = ModuleEditor()
-    editor.open_module(copy.deepcopy(basic_program))
-    return editor
-
-
-@pytest.fixture
-def intelligence(basic_program: Program) -> IntelligenceReadInterface:
-    intelligence = IntelligenceService()
-    intelligence.add_module(basic_program, Path("/fake/path.fl"))
-    return intelligence
-
-
-def test_move_function(
-    basic_program: Program,
-    editor: ModuleEditor,
-    intelligence: IntelligenceReadInterface,
-) -> None:
-    original = copy.deepcopy(basic_program)
-    expected = copy.deepcopy(basic_program)
-    expected.declarations[0].location = elements.Location(22, 57, 3, 100, 100)
-
-    uut = MoveElement(target=[1], x=22, y=57)
-
-    uut.resolve(intelligence, cast(Path, editor.get_path()))
-    editor.edit(uut)
-    actual = editor.get()
-
-    assert expected == actual
-
-    actual = uut.undo(actual)
-    assert original == actual
-
-
-def test_move_node(
-    basic_program: Program,
-    editor: ModuleEditor,
-    intelligence: IntelligenceReadInterface,
-) -> None:
-    original = copy.deepcopy(basic_program)
-    expected = copy.deepcopy(basic_program)
-    expected.declarations[1].nodes[2].location = elements.Location(
-        6, 90, 1, 5, 5
-    )
-
-    uut = MoveElement(target=[2, 3], x=6, y=90)
-
-    uut.resolve(intelligence, cast(Path, editor.get_path()))
-    editor.edit(uut)
-    actual = editor.get()
-
-    assert expected == actual
-
-    actual = uut.undo(actual)
-    assert original == actual
-
-
-@pytest.mark.parametrize(
-    "x, y",
-    [(3, 101), (102, 13), (96, 12), (53, 97), (-3, 48), (0, -13), (-3, -13)],
-)
-# limit=elements.Location(210, 10, 2, 100, 100)
-# element=elements.Location(2, 12, 1, 5, 5)
-def test_move_node_raises_if_out_of_function(
-    basic_program: Program,
-    editor: ModuleEditor,
-    intelligence: IntelligenceReadInterface,
-    x: int,
-    y: int,
-) -> None:
-    uut = MoveElement(target=[2, 3], x=x, y=y)
-
-    with pytest.raises(BadEdit):
-        uut.resolve(intelligence, cast(Path, editor.get_path()))
-        editor.edit(uut)
-
-
-@pytest.mark.parametrize(
-    "name, target, input, expected",
-    [
-        (
-            "resize_node",
-            [3, 3],
-            {"width": 16, "height": 5},
-            {"width": 16, "height": 5},
-        ),
-        (
-            "resize_function_decl",
-            [1],
-            {"width": 160, "height": 150},
-            {"width": 160, "height": 150},
-        ),
-        (
-            "resize_function_decl_stays_larger_than_children",
-            [2],
-            {"width": 12, "height": 15},
-            {"width": 20, "height": 17},
-        ),
-        (
-            "resize_node_clamps_to_x_limits",
-            [3, 3],
-            {"width": 100, "height": 5},
-            {"width": 98, "height": 5},
-        ),
-        (
-            "resize_node_clamps_to_y_limits",
-            [3, 3],
-            {"width": 14, "height": 150},
-            {"width": 14, "height": 98},
-        ),
-        (
-            "resize_node_clamps_to_min_limits",
-            [3, 3],
-            {"width": 3, "height": -6},
-            {"width": 4, "height": 4},
-        ),
-    ],
-    ids=lambda x: x if isinstance(x, str) else "",
-)
-def test_resize_element(
-    basic_program: Program,
-    editor: ModuleEditor,
-    intelligence: IntelligenceReadInterface,
-    target: QualifiedID,
-    input: dict[str, int],
-    expected: dict[str, int],
-    name: str,
-) -> None:
-    original = copy.deepcopy(basic_program)
-    expected_program = copy.deepcopy(basic_program)
-    e = find_element(target, expected_program)
-    e.location.width = expected["width"]
-    e.location.height = expected["height"]
-
-    uut = ResizeElement(
-        target=target, width=input["width"], height=input["height"]
-    )
-
-    uut.resolve(intelligence, cast(Path, editor.get_path()))
-    editor.edit(uut)
-    actual = editor.get()
-
-    assert expected_program == actual
-
-    actual = uut.undo(actual)
-    assert original == actual
-
-
-@pytest.mark.parametrize(
-    "name, target, input, expected",
-    [
-        (
-            "resize_node",
-            [3, 3],
-            {"width": 16, "height": 5, "x": 5, "y": 10},
-            {"width": 16, "height": 5, "x": 5, "y": 10},
-        ),
-        (
-            "resize_function_decl",
-            [1],
-            {"width": 160, "height": 150, "x": 20, "y": 30},
-            {"width": 160, "height": 150, "x": 20, "y": 30},
-        ),
-        (
-            "resize_function_decl_stays_larger_than_children",
-            [2],
-            {"width": 12, "height": 15, "x": 15, "y": 17},
-            {"width": 20, "height": 17, "x": 15, "y": 17},
-        ),
-        (
-            "resize_node_clamps_to_x_limits",
-            [3, 3],
-            {"width": 100, "height": 5, "x": 2, "y": 12},
-            {"width": 98, "height": 5, "x": 2, "y": 12},
-        ),
-        (
-            "resize_node_clamps_to_y_limits",
-            [3, 3],
-            {"width": 14, "height": 150, "x": 14, "y": 20},
-            {"width": 14, "height": 98, "x": 14, "y": 20},
-        ),
-        (
-            "resize_node_clamps_to_min_limits",
-            [3, 3],
-            {"width": 3, "height": -6, "x": 4, "y": 4},
-            {"width": 4, "height": 4, "x": 4, "y": 4},
-        ),
-    ],
-    ids=lambda x: x if isinstance(x, str) else "",
-)
-def test_resize_element_with_location(
-    basic_program: Program,
-    editor: ModuleEditor,
-    intelligence: IntelligenceReadInterface,
-    target: QualifiedID,
-    input: dict[str, int],
-    expected: dict[str, int],
-    name: str,
-) -> None:
-    original = copy.deepcopy(basic_program)
-    expected_program = copy.deepcopy(basic_program)
-    e = find_element(target, expected_program)
-    e.location.width = expected["width"]
-    e.location.height = expected["height"]
-    e.location.x = expected["x"]
-    e.location.y = expected["y"]
-
-    uut = ResizeElement(
-        target=target,
-        width=input["width"],
-        height=input["height"],
-        x=input["x"],
-        y=input["y"],
-    )
-
-    uut.resolve(intelligence, cast(Path, editor.get_path()))
-    editor.edit(uut)
-    actual = editor.get()
-
-    assert expected_program == actual
-
-    actual = uut.undo(actual)
-    assert original == actual
-
-
-def test_rename_function(
-    basic_program: Program,
-    editor: ModuleEditor,
-    intelligence: IntelligenceReadInterface,
-) -> None:
-    original = copy.deepcopy(basic_program)
-    expected = copy.deepcopy(basic_program)
-    expected.declarations[1].name = "baz"
-
-    uut = RenameDeclaration(target=[2], name="baz")
-
-    uut.resolve(intelligence, cast(Path, editor.get_path()))
-    editor.edit(uut)
-    actual = editor.get()
-
-    assert expected == actual
-
-    actual = uut.undo(actual)
-    assert original == actual
-
-
-def test_rename_declaration_raises_if_target_is_not_a_decl(
-    basic_program: Program,
-    editor: ModuleEditor,
-    intelligence: IntelligenceReadInterface,
-) -> None:
-    uut = RenameDeclaration(target=[2, 1], name="baz")
-
-    with pytest.raises(BadEdit, match="Only a declaration may be renamed"):
-        uut.resolve(intelligence, cast(Path, editor.get_path()))
-        editor.edit(uut)
-
-
-def test_edit_constant(
-    basic_program: Program,
-    editor: ModuleEditor,
-    intelligence: IntelligenceReadInterface,
-) -> None:
-    original = copy.deepcopy(basic_program)
-    expected = copy.deepcopy(basic_program)
-    assert isinstance(expected.declarations[1].nodes[1], elements.Constant)
-    expected.declarations[1].nodes[1].value = "-5.67"
-
-    uut = UpdateConstant(target=[2, 2], value="-5.67")
-    uut.resolve(intelligence, cast(Path, editor.get_path()))
-
-    editor.edit(uut)
-
-    actual = editor.get()
-
-    assert expected == actual
-
-    actual = uut.undo(actual)
-    assert original == actual
-
-
-def test_edit_constant_raises_if_not_constant(
-    editor: ModuleEditor, intelligence: IntelligenceReadInterface
-) -> None:
-    uut = UpdateConstant(target=[2, 1], value="-5.67")
-    with pytest.raises(BadEdit):
-        uut.resolve(intelligence, cast(Path, editor.get_path()))
-        editor.edit(uut)
-
-
-@pytest.mark.parametrize(
-    "op",
-    ["+", "-", "*", "/"],
-)
-def test_update_binary_operator(
-    basic_program: Program,
-    editor: ModuleEditor,
-    intelligence: IntelligenceReadInterface,
-    op: str,
-) -> None:
-    original = copy.deepcopy(basic_program)
-    expected = copy.deepcopy(basic_program)
-    assert isinstance(
-        expected.declarations[1].nodes[0], elements.BinaryOperator
-    )
-    expected.declarations[1].nodes[0].op = elements.Operator(op)
-
-    uut = UpdateOperator(target=[2, 1], value=op)
-    uut.resolve(intelligence, cast(Path, editor.get_path()))
-
-    editor.edit(uut)
-
-    actual = editor.get()
-
-    assert expected == actual
-
-    actual = uut.undo(actual)
-    assert original == actual
-
-
-@pytest.mark.parametrize(
-    "op",
-    ["+", "-", "*", "/"],
-)
-def test_update_unary_operator(
-    basic_program: Program,
-    editor: ModuleEditor,
-    intelligence: IntelligenceReadInterface,
-    op: str,
-) -> None:
-    original = copy.deepcopy(basic_program)
-    expected = copy.deepcopy(basic_program)
-    assert isinstance(expected.declarations[2].nodes[0], elements.UnaryOperator)
-    expected.declarations[2].nodes[0].op = elements.Operator(op)
-
-    uut = UpdateOperator(target=[3, 2], value=op)
-    uut.resolve(intelligence, cast(Path, editor.get_path()))
-
-    editor.edit(uut)
-
-    actual = editor.get()
-
-    assert expected == actual
-
-    actual = uut.undo(actual)
-    assert original == actual
-
-
-def test_update_operator_raises_if_not_operator(
-    editor: ModuleEditor, intelligence: IntelligenceReadInterface
-) -> None:
-    uut = UpdateOperator(target=[2, 2], value=elements.Operator.MINUS)
-    with pytest.raises(
-        BadEdit, match="Can only perform this update on an operator"
-    ):
-        uut.resolve(intelligence, cast(Path, editor.get_path()))
-        editor.edit(uut)
-
-
-def test_add_conduit_no_conflicts(
-    basic_program: Program,
-    editor: ModuleEditor,
-    intelligence: IntelligenceReadInterface,
-) -> None:
-    original = copy.deepcopy(basic_program)
-    expected = copy.deepcopy(basic_program)
-    new_conduit = elements.Conduit(
-        id=6,
-        input=2,
-        index=0,
-        children=[
-            elements.Conduit.Output(target=1, index=0),
-        ],
-    )
-    expected.declarations[1].conduits.append(new_conduit)
-
-    uut = AddConduit(target="input-2:1-0", source="output-2:2-0")
-    uut.resolve(intelligence, cast(Path, editor.get_path()))
-
-    editor.edit(uut)
-
-    actual = editor.get()
-
-    assert expected == actual
-
-    actual = uut.undo(actual)
-    assert original == actual
-
-
-def test_add_conduit_removes_duplicate_targets(
-    basic_program: Program,
-    editor: ModuleEditor,
-    intelligence: IntelligenceReadInterface,
-) -> None:
-    existing_conduit = elements.Conduit(
-        id=4,
-        input=3,
-        index=0,
-        children=[
-            elements.Conduit.Output(target=1, index=0),
-        ],
-    )
-    # This conduit targets the same output as the new conduit,
-    # so it should be overwritten
-    program = editor.get()
-    assert program is not None
-    program.declarations[1].conduits.append(existing_conduit)
-
-    original = copy.deepcopy(program)
-    expected = copy.deepcopy(basic_program)
-    new_conduit = elements.Conduit(
-        id=4,
-        input=2,
-        index=0,
-        children=[
-            elements.Conduit.Output(target=1, index=0),
-        ],
-    )
-    expected.declarations[1].conduits.append(new_conduit)
-
-    uut = AddConduit(target="input-2:1-0", source="output-2:2-0")
-    uut.resolve(intelligence, cast(Path, editor.get_path()))
-    editor.edit(uut)
-    actual = editor.get()
-
-    assert expected == actual
-
-    actual = uut.undo(actual)
-    assert original == actual
-
-
-@pytest.mark.parametrize(
-    "expected_node, input",
-    [
-        (
-            elements.Constant(
-                id=6,
-                location=elements.Location(2, 2, 0, 5, 5),
-                value="0.0",
-                flType=FlType.F64,
-            ),
-            AddNode(
-                parent=[2],
-                new_location=elements.Location(
-                    x=2, y=2, z=0, width=5, height=5
-                ),
-                params=ConstantParams(type=FlType.F64),
-            ),
-        ),
-        *(
-            (
-                elements.Constant(
-                    id=6,
-                    location=elements.Location(2, 2, 0, 5, 5),
-                    value="0",
-                    flType=FlType(elem),
-                ),
-                AddNode(
-                    parent=[2],
-                    new_location=elements.Location(
-                        x=2, y=2, z=0, width=5, height=5
-                    ),
-                    params=ConstantParams(type=FlType(elem)),
-                ),
-            )
-            for elem in ["I8", "I16", "I32", "I64", "U8", "U16", "U32", "U64"]
-        ),
-        (
-            elements.BinaryOperator(
-                id=6,
-                location=elements.Location(15, 2, 1, 5, 5),
-                op=elements.Operator.PLUS,
-            ),
-            AddNode(
-                parent=[2],
-                new_location=elements.Location(15, 2, 1, 5, 5),
-                params=OperatorParams(arity="binary"),
-            ),
-        ),
-        (
-            elements.UnaryOperator(
-                id=6,
-                location=elements.Location(2, 7, 0, 7, 7),
-                op=elements.Operator.PLUS,
-            ),
-            AddNode(
-                parent=[2],
-                new_location=elements.Location(2, 7, 0, 7, 7),
-                params=OperatorParams(arity="unary"),
-            ),
-        ),
-        # Constant with explicit value
-        (
-            elements.Constant(
-                id=6,
-                location=elements.Location(2, 2, 0, 5, 5),
-                value="42",
-                flType=FlType.I32,
-            ),
-            AddNode(
-                parent=[2],
-                new_location=elements.Location(2, 2, 0, 5, 5),
-                params=ConstantParams(type=FlType.I32, value="42"),
-            ),
-        ),
-        (
-            elements.Constant(
-                id=6,
-                location=elements.Location(2, 2, 0, 5, 5),
-                value="3.14",
-                flType=FlType.F64,
-            ),
-            AddNode(
-                parent=[2],
-                new_location=elements.Location(2, 2, 0, 5, 5),
-                params=ConstantParams(type=FlType.F64, value="3.14"),
-            ),
-        ),
-        # Operator with explicit op
-        (
-            elements.BinaryOperator(
-                id=6,
-                location=elements.Location(15, 2, 1, 5, 5),
-                op=elements.Operator.STAR,
-            ),
-            AddNode(
-                parent=[2],
-                new_location=elements.Location(15, 2, 1, 5, 5),
-                params=OperatorParams(arity="binary", op="*"),
-            ),
-        ),
-        (
-            elements.UnaryOperator(
-                id=6,
-                location=elements.Location(2, 7, 0, 7, 7),
-                op=elements.Operator.MINUS_MINUS,
-            ),
-            AddNode(
-                parent=[2],
-                new_location=elements.Location(2, 7, 0, 7, 7),
-                params=OperatorParams(arity="unary", op="--"),
-            ),
-        ),
-        # Operator with UNKNOWN op defaults to PLUS
-        (
-            elements.BinaryOperator(
-                id=6,
-                location=elements.Location(15, 2, 1, 5, 5),
-                op=elements.Operator.PLUS,
-            ),
-            AddNode(
-                parent=[2],
-                new_location=elements.Location(15, 2, 1, 5, 5),
-                params=OperatorParams(arity="binary", op=" "),
-            ),
-        ),
-        (
-            elements.Call(
-                id=6,
-                location=elements.Location(2, 7, 0, 7, 15),
-                target="xyzzy",
-                arguments=["a", "b"],
-                returns=True,
-            ),
-            AddNode(
-                parent=[2],
-                new_location=elements.Location(2, 7, 0, 7, 7),
-                params=CallParams(target="xyzzy"),
-            ),
-        ),
-        # Calling an unknown function adds its name and empty args/return
-        (
-            elements.Call(
-                id=6,
-                location=elements.Location(2, 7, 0, 7, 5),
-                target="unknown",
-                arguments=[],
-                returns=False,
-            ),
-            AddNode(
-                parent=[2],
-                new_location=elements.Location(2, 7, 0, 7, 7),
-                params=CallParams(target="unknown"),
-            ),
-        ),
-    ],
-)
-def test_add_node(
-    expected_node: elements.Node,
-    input: AddNode,
-    basic_program: Program,
-    editor: ModuleEditor,
-    intelligence: IntelligenceReadInterface,
-) -> None:
-    original = copy.deepcopy(basic_program)
-    expected = copy.deepcopy(basic_program)
-    expected.declarations[1].nodes.append(expected_node)
-
-    input.resolve(intelligence, cast(Path, editor.get_path()))
-    editor.edit(input)
-    actual = editor.get()
-
-    assert expected == actual
-
-    actual = input.undo(actual)
-    assert original == actual
-
-
-def test_add_node_missing_params_fails() -> None:
-    with pytest.raises(ValidationError):
-        AddNode(  # type: ignore[call-arg]
-            parent=[2],
-            new_location=elements.Location(2, 7, 0, 7, 7),
-        )
-
-
-def test_add_node_invalid_arity_fails() -> None:
-    with pytest.raises(ValidationError):
-        OperatorParams(arity="not_valid")  # type: ignore[arg-type]
-
-
-def test_add_call_node_without_target_raises(
-    basic_program: Program, editor: ModuleEditor
-) -> None:
-    """Adding a Call node with no target should raise BadEdit rather
-    than silently inserting a placeholder '???' target."""
-    uut = AddNode(
-        parent=[2],
-        new_location=elements.Location(2, 7, 0, 7, 7),
-        params=CallParams(),
-    )
-
-    with pytest.raises(BadEdit):
-        uut.do(basic_program)
-
-
-def test_add_node_with_invalid_op_fails(
-    basic_program: Program,
-    editor: ModuleEditor,
-    intelligence: IntelligenceReadInterface,
-) -> None:
-    uut = AddNode(
-        parent=[2],
-        new_location=elements.Location(2, 7, 0, 7, 7),
-        params=OperatorParams(arity="binary", op="invalid"),
-    )
-    with pytest.raises(BadEdit):
-        uut.resolve(intelligence, cast(Path, editor.get_path()))
-        editor.edit(uut)
-
-
-def test_add_call_node_can_add_builtin(
-    basic_program: Program,
-    editor: ModuleEditor,
-    intelligence: IntelligenceReadInterface,
-) -> None:
-    expected_node = elements.Call(
-        id=6,
-        location=elements.Location(2, 7, 0, 7, 10),
-        target="print",
-        arguments=["object"],
-        returns=False,
-    )
-    input = AddNode(
-        parent=[2],
-        new_location=elements.Location(2, 7, 0, 7, 7),
-        params=CallParams(target="print"),
-    )
-
-    original = copy.deepcopy(basic_program)
-    expected = copy.deepcopy(basic_program)
-    expected.declarations[1].nodes.append(expected_node)
-
-    input.resolve(intelligence, cast(Path, editor.get_path()))
-    editor.edit(input)
-    actual = editor.get()
-
-    assert expected == actual
-
-    actual = input.undo(actual)
-    assert original == actual
-
-
-def test_add_decl(
-    basic_program: Program,
-    editor: ModuleEditor,
-    intelligence: IntelligenceReadInterface,
-) -> None:
-    original = copy.deepcopy(basic_program)
-    expected = copy.deepcopy(basic_program)
-    expected.declarations.append(
-        elements.Function(
-            id=5,
-            name="new_function",
-            location=elements.Location(50, 50, 0, 200, 200),
-        )
-    )
-
-    uut = AddDecl(new_location=elements.Location(50, 50, 0, 200, 200))
-    uut.resolve(intelligence, cast(Path, editor.get_path()))
-    editor.edit(uut)
-    actual = editor.get()
-
-    assert expected == actual
-
-    actual = uut.undo(actual)
-    assert original == actual
-
-
-def test_add_decl_with_custom_name(
-    basic_program: Program,
-    editor: ModuleEditor,
-    intelligence: IntelligenceReadInterface,
-) -> None:
-    original = copy.deepcopy(basic_program)
-    expected = copy.deepcopy(basic_program)
-    expected.declarations.append(
-        elements.Function(
-            id=5,
-            name="my_func",
-            location=elements.Location(50, 50, 0, 200, 200),
-        )
-    )
-
-    uut = AddDecl(
-        new_location=elements.Location(50, 50, 0, 200, 200),
-        params=FunctionParams(name="my_func"),
-    )
-    uut.resolve(intelligence, cast(Path, editor.get_path()))
-    editor.edit(uut)
-    actual = editor.get()
-
-    assert expected == actual
-
-    actual = uut.undo(actual)
-    assert original == actual
-
-
-def test_remove_node(
-    basic_program: Program,
-    editor: ModuleEditor,
-    intelligence: IntelligenceReadInterface,
-) -> None:
-    original = copy.deepcopy(basic_program)
-    expected = copy.deepcopy(basic_program)
-    expected.declarations[1].nodes.pop(2)
-
-    uut = RemoveItem(target=[2, 3])
-    uut.resolve(intelligence, cast(Path, editor.get_path()))
-    editor.edit(uut)
-    actual = editor.get()
-
-    assert expected == actual
-
-    actual = uut.undo(actual)
-    assert original == actual
-
-
-def test_add_func_parameter(
-    basic_program: Program,
-    editor: ModuleEditor,
-    intelligence: IntelligenceReadInterface,
-) -> None:
-    original = copy.deepcopy(basic_program)
-    expected = copy.deepcopy(basic_program)
-    expected.declarations[0].inputs = [
-        elements.Parameter(name="new_param", id=1, flType=FlType.U8),
-    ]
-
-    uut = AddDeclInterface(
-        parent=[1],
-        flType=FlType.U8,
-        params=DeclParameterParams(name="new_param"),
-    )
-
-    uut.resolve(intelligence, cast(Path, editor.get_path()))
-    editor.edit(uut)
-    actual = editor.get()
-
-    assert actual is not None
-    assert actual == expected
-
-    actual = uut.undo(actual)
-    assert original == actual
-
-
-def test_add_multiple_func_parameter(
-    basic_program: Program, editor: ModuleEditor
-) -> None:
-    expected = copy.deepcopy(basic_program)
-    expected.declarations[0].inputs = [
-        elements.Parameter(name="new_param", id=1, flType=FlType.U8),
-        elements.Parameter(name="other_param", id=2, flType=FlType.U32),
-    ]
-
-    uut1 = AddDeclInterface(
-        parent=[1],
-        flType=FlType.U8,
-        params=DeclParameterParams(name="new_param"),
-    )
-    uut2 = AddDeclInterface(
-        parent=[1],
-        flType=FlType.U32,
-        params=DeclParameterParams(name="other_param"),
-    )
-
-    editor.edit(uut1)
-    editor.edit(uut2)
-    actual = editor.get()
-
-    assert actual is not None
-    assert actual == expected
-
-
-def test_add_decl_interface_raises_if_target_is_not_a_decl(
-    basic_program: Program,
-    editor: ModuleEditor,
-    intelligence: IntelligenceReadInterface,
-) -> None:
-    """AddDeclInterface should reject a parent that points to a node
-    inside a function rather than the function itself."""
-    # [2, 1] points to the BinaryOperator inside function bar, not bar.
-    uut = AddDeclInterface(
-        parent=[2, 1],
-        flType=FlType.U8,
-        params=DeclParameterParams(name="x"),
-    )
-
-    with pytest.raises(BadEdit):
-        uut.resolve(intelligence, cast(Path, editor.get_path()))
-        editor.edit(uut)
-
-
-def test_add_decl_interface_raises_if_decl_does_not_exist(
-    basic_program: Program,
-    editor: ModuleEditor,
-    intelligence: IntelligenceReadInterface,
-) -> None:
-    """AddDeclInterface should surface find_element's lookup error when
-    the parent ID does not resolve to any element."""
-    uut = AddDeclInterface(
-        parent=[9999],
-        flType=FlType.U8,
-        params=DeclParameterParams(name="x"),
-    )
-
-    with pytest.raises(elements.IdentifierError):
-        uut.resolve(intelligence, cast(Path, editor.get_path()))
-        editor.edit(uut)
-
-
-def test_add_parameter_with_empty_name_raises(
-    basic_program: Program,
-    editor: ModuleEditor,
-    intelligence: IntelligenceReadInterface,
-) -> None:
-    """A parameter must have a non-empty name."""
-    uut = AddDeclInterface(
-        parent=[1],
-        flType=FlType.U8,
-        params=DeclParameterParams(name=""),
-    )
-
-    with pytest.raises(BadEdit):
-        uut.resolve(intelligence, cast(Path, editor.get_path()))
-        editor.edit(uut)
-
-
-def test_add_func_return(
-    basic_program: Program,
-    editor: ModuleEditor,
-    intelligence: IntelligenceReadInterface,
-) -> None:
-    original = copy.deepcopy(basic_program)
-    expected = copy.deepcopy(basic_program)
-    expected.declarations[0].outputs = [
-        elements.Return(id=1, flType=FlType.U64)
-    ]
-
-    uut = AddDeclInterface(
-        parent=[1],
-        flType=FlType.U64,
-        params=DeclReturnParams(),
-    )
-
-    uut.resolve(intelligence, cast(Path, editor.get_path()))
-    editor.edit(uut)
-    actual = editor.get()
-
-    assert actual is not None
-    assert actual == expected
-
-    actual = uut.undo(actual)
-    assert original == actual
-
-
-def test_remove_node_with_conduits(
-    basic_program: Program,
-    editor: ModuleEditor,
-    intelligence: IntelligenceReadInterface,
-) -> None:
-    original = copy.deepcopy(basic_program)
-    expected = copy.deepcopy(basic_program)
-    expected.declarations[2].nodes.pop(1)
-    expected.declarations[2].conduits.pop()
-
-    uut = RemoveItem(target=[3, 3])
-    uut.resolve(intelligence, cast(Path, editor.get_path()))
-    editor.edit(uut)
-    actual = editor.get()
-
-    assert expected == actual
-
-    actual = uut.undo(actual)
-    assert original == actual
-
-
-def test_remove_function(
-    basic_program: Program,
-    editor: ModuleEditor,
-    intelligence: IntelligenceReadInterface,
-) -> None:
-    original = copy.deepcopy(basic_program)
-    expected = copy.deepcopy(basic_program)
-    expected.declarations.pop()
-
-    uut = RemoveItem(target=[4])
-    uut.resolve(intelligence, cast(Path, editor.get_path()))
-    editor.edit(uut)
-    actual = editor.get()
-
-    assert expected == actual
-
-    actual = uut.undo(actual)
-    assert original == actual
-
-
-def test_remove_conduit(
-    basic_program: Program,
-    editor: ModuleEditor,
-    intelligence: IntelligenceReadInterface,
-) -> None:
-    original = copy.deepcopy(basic_program)
-    expected = copy.deepcopy(basic_program)
-    expected.declarations[1].conduits.pop(0)
-
-    uut = RemoveItem(target=[2, 5])
-    uut.resolve(intelligence, cast(Path, editor.get_path()))
-    editor.edit(uut)
-    actual = editor.get()
-
-    assert expected == actual
-
-    actual = uut.undo(actual)
-    assert original == actual
-
-
-def test_remove_input(
-    basic_program: Program,
-    editor: ModuleEditor,
-    intelligence: IntelligenceReadInterface,
-) -> None:
-    basic_program.declarations[1].inputs.append(
-        elements.Parameter(name="x", id=10, flType=FlType.F64)
-    )
-    editor.open_module(copy.deepcopy(basic_program))
-
-    original = copy.deepcopy(basic_program)
-    expected = copy.deepcopy(basic_program)
-    expected.declarations[1].inputs = []
-
-    uut = RemoveItem(target=[2, 10])
-    uut.resolve(intelligence, cast(Path, editor.get_path()))
-    editor.edit(uut)
-    actual = editor.get()
-
-    assert expected == actual
-
-    actual = uut.undo(actual)
-    assert original == actual
-
-
-def test_remove_input_with_conduits(
-    basic_program: Program,
-    editor: ModuleEditor,
-    intelligence: IntelligenceReadInterface,
-) -> None:
-    basic_program.declarations[1].inputs.append(
-        elements.Parameter(name="x", id=10, flType=FlType.F64)
-    )
-    basic_program.declarations[1].conduits.append(
-        elements.Conduit(
-            id=11,
-            input=10,
-            children=[elements.Conduit.Output(target=1, index=0)],
-        )
-    )
-    editor.open_module(copy.deepcopy(basic_program))
-
-    original = copy.deepcopy(basic_program)
-    expected = copy.deepcopy(basic_program)
-    expected.declarations[1].inputs = []
-    expected.declarations[1].conduits = [
-        c for c in expected.declarations[1].conduits if c.id != 11
-    ]
-
-    uut = RemoveItem(target=[2, 10])
-    uut.resolve(intelligence, cast(Path, editor.get_path()))
-    editor.edit(uut)
-    actual = editor.get()
-
-    assert expected == actual
-
-    actual = uut.undo(actual)
-    assert original == actual
-
-
-def test_remove_output(
-    basic_program: Program,
-    editor: ModuleEditor,
-    intelligence: IntelligenceReadInterface,
-) -> None:
-    basic_program.declarations[1].outputs.append(
-        elements.Return(id=10, flType=FlType.F64)
-    )
-    editor.open_module(copy.deepcopy(basic_program))
-
-    original = copy.deepcopy(basic_program)
-    expected = copy.deepcopy(basic_program)
-    expected.declarations[1].outputs = []
-
-    uut = RemoveItem(target=[2, 10])
-    uut.resolve(intelligence, cast(Path, editor.get_path()))
-    editor.edit(uut)
-    actual = editor.get()
-
-    assert expected == actual
-
-    actual = uut.undo(actual)
-    assert original == actual
-
-
-def test_remove_output_with_conduits(
-    basic_program: Program,
-    editor: ModuleEditor,
-    intelligence: IntelligenceReadInterface,
-) -> None:
-    basic_program.declarations[1].outputs.append(
-        elements.Return(id=10, flType=FlType.F64)
-    )
-    basic_program.declarations[1].conduits.append(
-        elements.Conduit(
-            id=11,
-            input=1,
-            children=[elements.Conduit.Output(target=10, index=0)],
-        )
-    )
-    editor.open_module(copy.deepcopy(basic_program))
-
-    original = copy.deepcopy(basic_program)
-    expected = copy.deepcopy(basic_program)
-    expected.declarations[1].outputs = []
-    expected.declarations[1].conduits = [
-        c for c in expected.declarations[1].conduits if c.id != 11
-    ]
-
-    uut = RemoveItem(target=[2, 10])
-    uut.resolve(intelligence, cast(Path, editor.get_path()))
-    editor.edit(uut)
-    actual = editor.get()
-
-    assert expected == actual
-
-    actual = uut.undo(actual)
-    assert original == actual
-
-
-@pytest.fixture
-def program_with_comments(basic_program: Program) -> Program:
-    """basic_program with one top-level and one in-body comment preloaded."""
-    program = copy.deepcopy(basic_program)
-    program.annotations.append(
-        elements.Comment(
-            id=10,
-            location=elements.Location(5, 5, 0, 20, 10),
-            data="top-level comment",
-        )
-    )
-    program.declarations[1].annotations.append(
-        elements.Comment(
-            id=20,
-            location=elements.Location(3, 3, 0, 15, 8),
-            data="in-body comment",
-        )
-    )
-    return program
-
-
-@pytest.fixture
-def comments_editor(program_with_comments: Program) -> ModuleEditor:
-    editor = ModuleEditor()
-    editor.open_module(copy.deepcopy(program_with_comments))
-    return editor
-
-
-def test_add_comment_top_level_to_empty_program(
-    intelligence: IntelligenceReadInterface,
-) -> None:
-    program = Program()
-    editor = ModuleEditor()
-    editor.open_module(copy.deepcopy(program))
-
-    uut = AddComment(
-        parent=[],
-        new_location=elements.Location(1, 2, 0, 30, 15),
-        data="hello",
-    )
-    uut.resolve(intelligence, cast(Path, editor.get_path()))
-    editor.edit(uut)
-    actual = editor.get()
-    assert actual is not None
-    assert len(actual.annotations) == 1
-    assert actual.annotations[0].id == 1
-    assert actual.annotations[0].data == "hello"
-    assert actual.annotations[0].location == elements.Location(1, 2, 0, 30, 15)
-
-
-def test_add_comment_top_level_uses_next_decl_id(
-    basic_program: Program,
-    editor: ModuleEditor,
-    intelligence: IntelligenceReadInterface,
-) -> None:
-    original = copy.deepcopy(basic_program)
-    expected = copy.deepcopy(basic_program)
-    expected.annotations.append(
-        elements.Comment(
-            id=5,
-            location=elements.Location(7, 8, 0, 25, 12),
-            data="note",
-        )
-    )
-
-    uut = AddComment(
-        parent=[],
-        new_location=elements.Location(7, 8, 0, 25, 12),
-        data="note",
-    )
-    uut.resolve(intelligence, cast(Path, editor.get_path()))
-    editor.edit(uut)
-    actual = editor.get()
-
-    assert expected == actual
-
-    actual = uut.undo(actual)
-    assert original == actual
-
-
-def test_add_comment_in_body_uses_function_local_next_id(
-    basic_program: Program,
-    editor: ModuleEditor,
-    intelligence: IntelligenceReadInterface,
-) -> None:
-    original = copy.deepcopy(basic_program)
-    expected = copy.deepcopy(basic_program)
-    expected.declarations[1].annotations.append(
-        elements.Comment(
-            id=6,
-            location=elements.Location(4, 4, 0, 10, 6),
-            data="inside",
-        )
-    )
-
-    uut = AddComment(
-        parent=[2],
-        new_location=elements.Location(4, 4, 0, 10, 6),
-        data="inside",
-    )
-    uut.resolve(intelligence, cast(Path, editor.get_path()))
-    editor.edit(uut)
-    actual = editor.get()
-
-    assert expected == actual
-
-    actual = uut.undo(actual)
-    assert original == actual
-
-
-def test_add_comment_undo_via_editor(
-    basic_program: Program,
-    editor: ModuleEditor,
-    intelligence: IntelligenceReadInterface,
-) -> None:
-    original = copy.deepcopy(basic_program)
-
-    uut = AddComment(
-        parent=[],
-        new_location=elements.Location(1, 1, 0, 10, 5),
-        data="x",
-    )
-    uut.resolve(intelligence, cast(Path, editor.get_path()))
-    editor.edit(uut)
-    editor.undo()
-    actual = editor.get()
-
-    assert actual == original
-
-
-def test_add_comment_redo_via_editor(
-    basic_program: Program,
-    editor: ModuleEditor,
-    intelligence: IntelligenceReadInterface,
-) -> None:
-    expected = copy.deepcopy(basic_program)
-    expected.annotations.append(
-        elements.Comment(
-            id=5,
-            location=elements.Location(1, 1, 0, 10, 5),
-            data="x",
-        )
-    )
-
-    uut = AddComment(
-        parent=[],
-        new_location=elements.Location(1, 1, 0, 10, 5),
-        data="x",
-    )
-    uut.resolve(intelligence, cast(Path, editor.get_path()))
-    editor.edit(uut)
-    editor.undo()
-    editor.redo()
-    actual = editor.get()
-
-    assert actual == expected
-
-
-def test_add_comment_in_body_then_undo_redo(
-    basic_program: Program,
-    editor: ModuleEditor,
-    intelligence: IntelligenceReadInterface,
-) -> None:
-    original = copy.deepcopy(basic_program)
-    expected = copy.deepcopy(basic_program)
-    expected.declarations[1].annotations.append(
-        elements.Comment(
-            id=6,
-            location=elements.Location(2, 2, 0, 8, 4),
-            data="hi",
-        )
-    )
-
-    uut = AddComment(
-        parent=[2],
-        new_location=elements.Location(2, 2, 0, 8, 4),
-        data="hi",
-    )
-    uut.resolve(intelligence, cast(Path, editor.get_path()))
-    editor.edit(uut)
-    editor.undo()
-    assert editor.get() == original
-    editor.redo()
-    assert editor.get() == expected
-
-
-def test_update_comment_top_level(
-    program_with_comments: Program,
-    comments_editor: ModuleEditor,
-    intelligence: IntelligenceReadInterface,
-) -> None:
-    original = copy.deepcopy(program_with_comments)
-    expected = copy.deepcopy(program_with_comments)
-    expected.annotations[0].data = "edited top-level"
-
-    uut = UpdateComment(target=[10], data="edited top-level")
-    uut.resolve(intelligence, cast(Path, comments_editor.get_path()))
-    comments_editor.edit(uut)
-    actual = comments_editor.get()
-
-    assert expected == actual
-
-    actual = uut.undo(actual)
-    assert original == actual
-
-
-def test_update_comment_in_body(
-    program_with_comments: Program,
-    comments_editor: ModuleEditor,
-    intelligence: IntelligenceReadInterface,
-) -> None:
-    original = copy.deepcopy(program_with_comments)
-    expected = copy.deepcopy(program_with_comments)
-    expected.declarations[1].annotations[0].data = "edited body"
-
-    uut = UpdateComment(target=[2, 20], data="edited body")
-    uut.resolve(intelligence, cast(Path, comments_editor.get_path()))
-    comments_editor.edit(uut)
-    actual = comments_editor.get()
-
-    assert expected == actual
-
-    actual = uut.undo(actual)
-    assert original == actual
-
-
-def test_update_comment_redo_via_editor(
-    program_with_comments: Program,
-    comments_editor: ModuleEditor,
-    intelligence: IntelligenceReadInterface,
-) -> None:
-    expected = copy.deepcopy(program_with_comments)
-    expected.annotations[0].data = "redo me"
-
-    uut = UpdateComment(target=[10], data="redo me")
-    uut.resolve(intelligence, cast(Path, comments_editor.get_path()))
-    comments_editor.edit(uut)
-    comments_editor.undo()
-    comments_editor.redo()
-
-    assert comments_editor.get() == expected
-
-
-def test_update_comment_raises_if_not_comment(
-    editor: ModuleEditor,
-    intelligence: IntelligenceReadInterface,
-) -> None:
-    uut = UpdateComment(target=[2, 1], data="nope")
-    with pytest.raises(BadEdit):
-        uut.resolve(intelligence, cast(Path, editor.get_path()))
-        editor.edit(uut)
-
-
-def test_move_top_level_comment(
-    program_with_comments: Program,
-    comments_editor: ModuleEditor,
-    intelligence: IntelligenceReadInterface,
-) -> None:
-    original = copy.deepcopy(program_with_comments)
-    expected = copy.deepcopy(program_with_comments)
-    expected.annotations[0].location.x = 50
-    expected.annotations[0].location.y = 60
-
-    uut = MoveElement(target=[10], x=50, y=60)
-    uut.resolve(intelligence, cast(Path, comments_editor.get_path()))
-    comments_editor.edit(uut)
-    actual = comments_editor.get()
-
-    assert expected == actual
-
-    actual = uut.undo(actual)
-    assert original == actual
-
-
-def test_resize_in_body_comment(
-    program_with_comments: Program,
-    comments_editor: ModuleEditor,
-    intelligence: IntelligenceReadInterface,
-) -> None:
-    original = copy.deepcopy(program_with_comments)
-    expected = copy.deepcopy(program_with_comments)
-    expected.declarations[1].annotations[0].location.width = 25
-    expected.declarations[1].annotations[0].location.height = 12
-
-    uut = ResizeElement(target=[2, 20], width=25, height=12)
-    uut.resolve(intelligence, cast(Path, comments_editor.get_path()))
-    comments_editor.edit(uut)
-    actual = comments_editor.get()
-
-    assert expected == actual
-
-    actual = uut.undo(actual)
-    assert original == actual
-
-
-def test_remove_top_level_comment(
-    program_with_comments: Program,
-    comments_editor: ModuleEditor,
-    intelligence: IntelligenceReadInterface,
-) -> None:
-    original = copy.deepcopy(program_with_comments)
-    expected = copy.deepcopy(program_with_comments)
-    expected.annotations.pop(0)
-
-    uut = RemoveItem(target=[10])
-    uut.resolve(intelligence, cast(Path, comments_editor.get_path()))
-    comments_editor.edit(uut)
-    actual = comments_editor.get()
-
-    assert expected == actual
-
-    actual = uut.undo(actual)
-    assert original == actual
-
-
-def test_remove_in_body_comment(
-    program_with_comments: Program,
-    comments_editor: ModuleEditor,
-    intelligence: IntelligenceReadInterface,
-) -> None:
-    original = copy.deepcopy(program_with_comments)
-    expected = copy.deepcopy(program_with_comments)
-    expected.declarations[1].annotations.pop(0)
-
-    uut = RemoveItem(target=[2, 20])
-    comments_editor.edit(uut)
-    actual = comments_editor.get()
-
-    assert expected == actual
-
-    actual = uut.undo(actual)
-    assert original == actual
-
-
-def test_update_func_param(
-    basic_program: Program,
-    editor: ModuleEditor,
-    intelligence: IntelligenceReadInterface,
-) -> None:
-    undone = copy.deepcopy(basic_program)
-    done = copy.deepcopy(basic_program)
-    done.declarations[3].inputs[0].flType = FlType.I64
-
-    input = UpdateFuncParam(target=[4], index=0, type=FlType.I64)
-
-    editor.edit(input)
-    actual = editor.get()
-
-    assert done == actual
-
-    editor.undo()
-    actual = editor.get()
-    assert undone == actual
